feat(mining): remove legacy pattern detector, TaskMiner only
The newTaskMinerEnabled toggle (default off) still selected the legacy
PatternDetector at startup; this is the real cutover. TaskMiner is now
the only analyzer.

- delete services/pattern-detector/, PatternRepository, legacy Patterns
  view (IPC, preload, types, renderer components) and the dev toggle
- move shared helpers/tools into services/task-miner/
- port the Developer > Tasks golden builder to the sightings table
- CLI: back stats patternCount with clusters; fix test seed broken by
  the recipe-steps schema bump (missing steps/variables)

Legacy patterns/pattern_sightings/pattern_detection_runs tables stay in
the DB (no drop migration); patternDetection* setting keys unchanged.
First launch on previously-default installs starts the ~60-day backfill.
