+++ src/main/eval/task-replay.ts
@@ -6,7 +6,7 @@ import { StorageService } from '../storage'
 import { applyMigrations } from '../storage/migrator'
 import { deleteDbFiles } from '../storage/test-utils'
 import type { StoredActivity } from '../storage/types'
-import { getDayBoundaries } from '../services/pattern-detector/helpers'
+import { getDayBoundaries } from '../services/task-miner/helpers'
 import { runDetection } from '../services/task-miner/run-detection'
 import { DEFAULT_MINER_CONFIG } from '@main/services/task-miner/types'
 import type { EmbeddingService } from '../processor/embedding'
