+++ src/main/services/task-miner/tools.ts
