+++ src/main/settings/capture-settings-manager.ts
@@ -162,7 +162,6 @@ const DEFAULTS: CaptureSettings = {
   patternDetectionModel: OPENROUTER_DEFAULTS.patternDetectionModel,
   modelsByVendor: {},
   patternDetectionEnabled: true,
-  newTaskMinerEnabled: false,
   uploadDetailLevel: 'off',
 }
