+++ src/main/storage/index.ts
@@ -5,7 +5,6 @@ import * as path from 'path'
 import { getDefaultDbPath } from '@main/utils/paths'
 import log from '@main/utils/logger'
 import { ActivityRepository } from './activity-repository'
-import { PatternRepository } from './pattern-repository'
 import { SightingRepository } from './sighting-repository'
 import { ClusterRepository } from './cluster-repository'
 import { MiningDayRepository } from './mining-day-repository'
@@ -13,13 +12,11 @@ import { UploadRunRepository } from './upload-run-repository'
 import { UserContextRepository } from './user-context-repository'
 
 export { ActivityRepository } from './activity-repository'
-export { PatternRepository } from './pattern-repository'
 export { SightingRepository } from './sighting-repository'
 export { ClusterRepository } from './cluster-repository'
 export { MiningDayRepository } from './mining-day-repository'
 export { UploadRunRepository } from './upload-run-repository'
 export { UserContextRepository } from './user-context-repository'
-export type { Pattern, PatternSighting, PatternWithStats } from './pattern-repository'
 export type { Sighting } from './sighting-repository'
 export type { Cluster, ClusterWithStats } from './cluster-repository'
 export type { MiningDay, MiningDayStatus } from './mining-day-repository'
@@ -72,7 +69,6 @@ export class StorageService {
   private dbPath: string
   private db: Database.Database | null = null
   readonly activities: ActivityRepository
-  readonly patterns: PatternRepository
   readonly sightings: SightingRepository
   readonly clusters: ClusterRepository
   readonly miningDays: MiningDayRepository
@@ -101,7 +97,6 @@ export class StorageService {
 
       this.db = db
       this.activities = new ActivityRepository(db)
-      this.patterns = new PatternRepository(db)
       this.sightings = new SightingRepository(db)
       this.clusters = new ClusterRepository(db)
       this.miningDays = new MiningDayRepository(db)
