--- src/main/storage/pattern-repository.ts
@@ -1,416 +1,0 @@
-import type Database from 'better-sqlite3'
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-export interface Pattern {
-  id: string
-  name: string
-  description: string
-  apps: string[] // JSON array in DB
-  automationIdea: string
-  createdAt: number
-  rejectedAt: number | null
-  promptCopiedAt: number | null
-  approvedAt: number | null
-  completedAt: number | null
-}
-
-export interface PatternSighting {
-  id: string
-  patternId: string
-  detectedAt: number
-  runId: string
-  evidence: string
-  activityIds: string[] // JSON array in DB
-  confidence: number
-  durationEstimateMin: number | null
-}
-
-/** Pattern with derived sighting stats (computed via JOIN). */
-export interface PatternWithStats extends Pattern {
-  sightingCount: number
-  lastSeenAt: number | null
-  lastConfidence: number | null
-  estimatedHoursPerWeek: number | null
-}
-
-export interface PatternDetail {
-  pattern: PatternWithStats
-  sightings: PatternSighting[]
-}
-
-/** A sighting joined with its parent pattern's name + apps (for flat listing). */
-export interface PatternSightingWithPattern extends PatternSighting {
-  patternName: string
-  patternApps: string[]
-}
-
-// ---------------------------------------------------------------------------
-// Scoring
-// ---------------------------------------------------------------------------
-
-const DAY_MS = 86_400_000
-
-function recencyDecay(lastSeenAt: number | null, now: number): number {
-  if (lastSeenAt === null) return 0
-  const ageDays = (now - lastSeenAt) / DAY_MS
-  if (ageDays < 7) return 1
-  if (ageDays < 30) return 0.5
-  return 0.2
-}
-
-export function computeScore(
-  input: {
-    estimatedHoursPerWeek: number | null
-    sightingCount: number
-    lastConfidence: number | null
-    lastSeenAt: number | null
-  },
-  now: number = Date.now(),
-): number {
-  const hours = input.estimatedHoursPerWeek ?? 0
-  if (hours <= 0) return 0
-  const recurrence = Math.log2(1 + input.sightingCount)
-  const confidence = input.lastConfidence ?? 0.5
-  const recency = recencyDecay(input.lastSeenAt, now)
-  return hours * recurrence * confidence * recency
-}
-
-// ---------------------------------------------------------------------------
-// Repository
-// ---------------------------------------------------------------------------
-
-interface CountRow {
-  readonly count: number
-}
-
-interface LastRunRow {
-  readonly latest: number | null
-}
-
-export class PatternRepository {
-  constructor(private readonly db: Database.Database) {}
-
-  addPattern(pattern: Pattern): void {
-    this.db
-      .prepare(
-        `INSERT OR IGNORE INTO patterns (id, name, description, apps, automation_idea, created_at)
-         VALUES (?, ?, ?, ?, ?, ?)`,
-      )
-      .run(
-        pattern.id,
-        pattern.name,
-        pattern.description,
-        JSON.stringify(pattern.apps),
-        pattern.automationIdea,
-        pattern.createdAt,
-      )
-  }
-
-  getPatternById(id: string): PatternWithStats | null {
-    const row = this.db
-      .prepare(
-        `SELECT p.*,
-                COUNT(s.id) AS sighting_count,
-                MAX(s.detected_at) AS last_seen_at,
-                MIN(s.detected_at) AS first_seen_at,
-                AVG(s.duration_estimate_min) AS avg_duration_min,
-                (SELECT confidence FROM pattern_sightings WHERE pattern_id = p.id ORDER BY detected_at DESC LIMIT 1) AS last_confidence
-         FROM patterns p
-         LEFT JOIN pattern_sightings s ON s.pattern_id = p.id
-         WHERE p.id = ?
-         GROUP BY p.id`,
-      )
-      .get(id) as Record<string, unknown> | undefined
-
-    return row ? this.rowToPatternWithStats(row) : null
-  }
-
-  getAllPatterns(): PatternWithStats[] {
-    const rows = this.db
-      .prepare(
-        `SELECT p.*,
-                COUNT(s.id) AS sighting_count,
-                MAX(s.detected_at) AS last_seen_at,
-                MIN(s.detected_at) AS first_seen_at,
-                AVG(s.duration_estimate_min) AS avg_duration_min,
-                (SELECT confidence FROM pattern_sightings WHERE pattern_id = p.id ORDER BY detected_at DESC LIMIT 1) AS last_confidence
-         FROM patterns p
-         LEFT JOIN pattern_sightings s ON s.pattern_id = p.id
-         WHERE p.rejected_at IS NULL
-         GROUP BY p.id`,
-      )
-      .all() as Record<string, unknown>[]
-
-    const now = Date.now()
-    const patterns = rows.map((row) => this.rowToPatternWithStats(row))
-    const scores = new Map(patterns.map((p) => [p.id, computeScore(p, now)]))
-    return patterns.sort((a, b) => {
-      const aActive = a.completedAt === null ? 1 : 0
-      const bActive = b.completedAt === null ? 1 : 0
-      if (aActive !== bActive) return bActive - aActive
-      const sa = scores.get(a.id) ?? 0
-      const sb = scores.get(b.id) ?? 0
-      if (sb !== sa) return sb - sa
-      if (b.sightingCount !== a.sightingCount) return b.sightingCount - a.sightingCount
-      return (b.lastSeenAt ?? 0) - (a.lastSeenAt ?? 0)
-    })
-  }
-
-  getPatternDetail(id: string): PatternDetail | null {
-    const pattern = this.getPatternById(id)
-    if (!pattern) return null
-    const sightings = this.getSightingsForPattern(id, 50)
-    return { pattern, sightings }
-  }
-
-  getRejectedPatterns(limit = 3): PatternWithStats[] {
-    const rows = this.db
-      .prepare(
-        `SELECT p.*,
-                COUNT(s.id) AS sighting_count,
-                MAX(s.detected_at) AS last_seen_at,
-                MIN(s.detected_at) AS first_seen_at,
-                AVG(s.duration_estimate_min) AS avg_duration_min,
-                (SELECT confidence FROM pattern_sightings WHERE pattern_id = p.id ORDER BY detected_at DESC LIMIT 1) AS last_confidence
-         FROM patterns p
-         LEFT JOIN pattern_sightings s ON s.pattern_id = p.id
-         WHERE p.rejected_at IS NOT NULL
-         GROUP BY p.id
-         ORDER BY sighting_count DESC
-         LIMIT ?`,
-      )
-      .all(limit) as Record<string, unknown>[]
-
-    return rows.map((row) => this.rowToPatternWithStats(row))
-  }
-
-  searchPatterns(query: string): PatternWithStats[] {
-    const like = `%${query}%`
-    const rows = this.db
-      .prepare(
-        `SELECT p.*,
-                COUNT(s.id) AS sighting_count,
-                MAX(s.detected_at) AS last_seen_at,
-                MIN(s.detected_at) AS first_seen_at,
-                AVG(s.duration_estimate_min) AS avg_duration_min,
-                (SELECT confidence FROM pattern_sightings WHERE pattern_id = p.id ORDER BY detected_at DESC LIMIT 1) AS last_confidence
-         FROM patterns p
-         LEFT JOIN pattern_sightings s ON s.pattern_id = p.id
-         WHERE p.rejected_at IS NULL
-           AND (p.name LIKE ? OR p.description LIKE ? OR p.apps LIKE ?)
-         GROUP BY p.id
-         ORDER BY sighting_count DESC`,
-      )
-      .all(like, like, like) as Record<string, unknown>[]
-
-    return rows.map((row) => this.rowToPatternWithStats(row))
-  }
-
-  patternCount(): number {
-    const row = this.db
-      .prepare('SELECT COUNT(*) as count FROM patterns WHERE rejected_at IS NULL')
-      .get() as CountRow
-    return row.count
-  }
-
-  // -- Status updates --
-
-  approvePattern(id: string): void {
-    this.db.prepare(`UPDATE patterns SET approved_at = ? WHERE id = ?`).run(Date.now(), id)
-  }
-
-  rejectPattern(id: string): void {
-    this.db.prepare(`UPDATE patterns SET rejected_at = ? WHERE id = ?`).run(Date.now(), id)
-  }
-
-  completePattern(id: string): void {
-    this.db.prepare(`UPDATE patterns SET completed_at = ? WHERE id = ?`).run(Date.now(), id)
-  }
-
-  uncompletePattern(id: string): void {
-    this.db.prepare(`UPDATE patterns SET completed_at = NULL WHERE id = ?`).run(id)
-  }
-
-  updatePattern(
-    id: string,
-    fields: { name?: string; description?: string; apps?: string[]; automationIdea?: string },
-  ): void {
-    const sets: string[] = []
-    const values: unknown[] = []
-    if (fields.name) {
-      sets.push('name = ?')
-      values.push(fields.name)
-    }
-    if (fields.description) {
-      sets.push('description = ?')
-      values.push(fields.description)
-    }
-    if (fields.apps) {
-      sets.push('apps = ?')
-      values.push(JSON.stringify(fields.apps))
-    }
-    if (fields.automationIdea) {
-      sets.push('automation_idea = ?')
-      values.push(fields.automationIdea)
-    }
-    if (sets.length === 0) return
-    values.push(id)
-    this.db.prepare(`UPDATE patterns SET ${sets.join(', ')} WHERE id = ?`).run(...values)
-  }
-
-  markPromptCopied(id: string): void {
-    this.db.prepare(`UPDATE patterns SET prompt_copied_at = ? WHERE id = ?`).run(Date.now(), id)
-  }
-
-  // -- Sightings --
-
-  addSighting(sighting: PatternSighting): void {
-    this.db
-      .prepare(
-        `INSERT INTO pattern_sightings (id, pattern_id, detected_at, run_id, evidence, activity_ids, confidence, duration_estimate_min)
-         VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
-      )
-      .run(
-        sighting.id,
-        sighting.patternId,
-        sighting.detectedAt,
-        sighting.runId,
-        sighting.evidence,
-        JSON.stringify(sighting.activityIds),
-        sighting.confidence,
-        sighting.durationEstimateMin,
-      )
-  }
-
-  getSightingsForPattern(patternId: string, limit = 20): PatternSighting[] {
-    const rows = this.db
-      .prepare(
-        `SELECT * FROM pattern_sightings WHERE pattern_id = ? ORDER BY detected_at DESC LIMIT ?`,
-      )
-      .all(patternId, limit) as Record<string, unknown>[]
-
-    return rows.map((row) => this.rowToSighting(row))
-  }
-
-  /** All sightings across patterns, newest first, with parent name + apps. */
-  getAllSightings(limit = 200): PatternSightingWithPattern[] {
-    const rows = this.db
-      .prepare(
-        `SELECT s.*, p.name AS pattern_name, p.apps AS pattern_apps
-         FROM pattern_sightings s
-         JOIN patterns p ON p.id = s.pattern_id
-         ORDER BY s.detected_at DESC
-         LIMIT ?`,
-      )
-      .all(limit) as Record<string, unknown>[]
-
-    return rows.map((row) => ({
-      ...this.rowToSighting(row),
-      patternName: row.pattern_name as string,
-      patternApps: JSON.parse((row.pattern_apps as string) || '[]') as string[],
-    }))
-  }
-
-  getSightingsByRunId(runId: string): PatternSighting[] {
-    const rows = this.db
-      .prepare(`SELECT * FROM pattern_sightings WHERE run_id = ? ORDER BY detected_at DESC`)
-      .all(runId) as Record<string, unknown>[]
-
-    return rows.map((row) => this.rowToSighting(row))
-  }
-
-  getLastRunTimestamp(): number | null {
-    const row = this.db
-      .prepare('SELECT MAX(ran_at) AS latest FROM pattern_detection_runs')
-      .get() as LastRunRow
-    return row.latest ?? null
-  }
-
-  recordRun(runId: string, findingsCount: number): void {
-    this.db
-      .prepare('INSERT INTO pattern_detection_runs (id, ran_at, findings_count) VALUES (?, ?, ?)')
-      .run(runId, Date.now(), findingsCount)
-  }
-
-  // -- Cleanup --
-
-  /**
-   * Delete sightings older than `maxAgeDays`. If a pattern has no remaining
-   * sightings after pruning, delete the pattern too. Returns counts.
-   */
-  pruneStale(maxAgeDays = 30): { sightings: number; patterns: number } {
-    const cutoff = Date.now() - maxAgeDays * 24 * 60 * 60 * 1000
-
-    const sightings = this.db
-      .prepare('DELETE FROM pattern_sightings WHERE detected_at < ?')
-      .run(cutoff).changes
-
-    const patterns = this.db
-      .prepare(
-        `DELETE FROM patterns WHERE id NOT IN (
-           SELECT DISTINCT pattern_id FROM pattern_sightings
-         )`,
-      )
-      .run().changes
-
-    return { sightings, patterns }
-  }
-
-  // -- Private helpers --
-
-  private rowToPatternWithStats(row: Record<string, unknown>): PatternWithStats {
-    const sightingCount = (row.sighting_count as number) || 0
-    const avgDurationMin = (row.avg_duration_min as number) ?? null
-    const firstSeenAt = (row.first_seen_at as number) ?? null
-    const lastSeenAt = (row.last_seen_at as number) ?? null
-
-    let estimatedHoursPerWeek: number | null = null
-    if (
-      avgDurationMin !== null &&
-      sightingCount >= 2 &&
-      firstSeenAt !== null &&
-      lastSeenAt !== null
-    ) {
-      const spanDays = (lastSeenAt - firstSeenAt) / DAY_MS
-      if (spanDays >= 1) {
-        const frequencyPerWeek = (sightingCount / spanDays) * 7
-        estimatedHoursPerWeek = Math.round(((frequencyPerWeek * avgDurationMin) / 60) * 10) / 10
-      }
-    }
-
-    const lastConfidence = (row.last_confidence as number) ?? null
-
-    return {
-      id: row.id as string,
-      name: row.name as string,
-      description: row.description as string,
-      apps: JSON.parse((row.apps as string) || '[]') as string[],
-      automationIdea: row.automation_idea as string,
-      createdAt: row.created_at as number,
-      rejectedAt: (row.rejected_at as number) ?? null,
-      promptCopiedAt: (row.prompt_copied_at as number) ?? null,
-      approvedAt: (row.approved_at as number) ?? null,
-      completedAt: (row.completed_at as number) ?? null,
-      sightingCount,
-      lastSeenAt,
-      lastConfidence,
-      estimatedHoursPerWeek,
-    }
-  }
-
-  private rowToSighting(row: Record<string, unknown>): PatternSighting {
-    return {
-      id: row.id as string,
-      patternId: row.pattern_id as string,
-      detectedAt: row.detected_at as number,
-      runId: row.run_id as string,
-      evidence: row.evidence as string,
-      activityIds: JSON.parse((row.activity_ids as string) || '[]') as string[],
-      confidence: row.confidence as number,
-      durationEstimateMin: (row.duration_estimate_min as number) ?? null,
-    }
-  }
-}
